requirements.txt
requests
gitpython
tqdm
py-polynomial
